services issues fixed

+++ DirtyToCleanCodeExample/Service/Example1/UserService.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿
 using Data;
 using Data.Example1;
 using Microsoft.AspNetCore.Http;
@@ -6,17 +6,26 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Service.Example1
 {
-	public class UserService
-	{
+    public class UserService
+    {
         private readonly DataContext _dbContext;
+        private readonly IHttpContextAccessor _httpContextAccessor;
+
         public UserService(IHttpContextAccessor httpContextAccessor, DataContext dbContext)
         {
+            _httpContextAccessor = httpContextAccessor;
             _dbContext = dbContext;
         }
 
         public async Task<string> Update(int userId, string newEmail)
         {
-            var user = await _dbContext.Users.AsNoTracking().FirstOrDefaultAsync(x => x.Id == userId);
+            var user = await _dbContext.Users.FirstOrDefaultAsync(x => x.Id == userId);
+
+            //Null kontrolü yapılmalıydı
+            if (user == null)
+            {
+                return "User not found.";
+            }
 
             user.Email = newEmail;
             user.UpdatedDate = DateTime.Now;
@@ -26,29 +35,30 @@ namespace Service.Example1
         }
 
         public async Task<string> Delete(int userId)
-        {
-            var user = await _dbContext.Users.FirstOrDefaultAsync(x => x.Id == userId);
+        {   
+            //
+            var user = await _dbContext.Users
+                .Include(u => u.Posts)//postları silmek için direkt olarak include ile çekmeliyiz
+                .FirstOrDefaultAsync(x => x.Id == userId);
+            
+            //null kontrolü 
+            if (user == null)
+            {
+                return "User not found.";
+            }
 
             _dbContext.Users.Remove(user);
             await _dbContext.SaveChangesAsync();
 
-            await DeletePost(userId);
+            await DeletePosts(user.Posts);
 
             return "User deleted successfully.";
         }
 
-        public async Task<bool> DeletePost(int userId)
+        private async Task<bool> DeletePosts(IEnumerable<Post> posts)
         {
-            var userPosts = _dbContext.Posts.Where(x => x.UserId == userId).ToList();
-            foreach (var post in userPosts)
-            {
-               _dbContext.Posts.Remove(post);
-               _dbContext.SaveChanges();
-            }
-
-            return true;
+            _dbContext.Posts.RemoveRange(posts); //postları range olarak silebiliriz
+            return await _dbContext.SaveChangesAsync() == 1;
         }
-
     }
 }
-
